Got all form fields writing/updating to db
Users can now enter and edit their details, which will update in the db.  Also got the passwords in the register form to check if they match. I made loads of routes in the dashboard for the edit options.  I feel there's probably an easier way to do this, maybe just one route/function, but this is the only way I could get it working for now.