Difficult to run locally
Trying to test the app on azure so some code had to change to make that happen (connection string), so now the local db won't connect.  Geting an error on azure when I try to upload a file (ENOENT: no such file or directory, open 'D:\home\site\wwwroot\bin\public\uploads\1464293582347_Cian'sConfirmation.jpg').  Not sure what to do with this yet.